seo-audit-agent/requirements.txt
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-analytics-data==0.18.7
google-ads==25.0.0
requests==2.32.3
pydantic==2.7.1
pandas==2.2.2
scikit-learn==1.5.0
numpy==1.26.4
tqdm==4.66.4
python-dotenv==1.0.1
tenacity==8.3.0
